# src/main/resources/api/api.yml
openapi: 3.0.1
info:
  title: Order API
  description: API pour gérer les commandes
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:

  /internal/api/v1/orders:
    get:
      tags:
        - "Orders"
      operationId: "getAllOrders"
      summary: Récupérer tous les commandes
      responses:
        '200':
          description: Succès. Renvoie la liste de tous les commandes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Mauvaise requête
        '404':
          description: Commandes non retrouvables
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - "Orders"
      operationId: "createOrder"
      summary: Ajoute une nouveau commande
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'

      responses:
        '201':
          description: Commande ajoutée avec succès
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Mauvaise requête.
        '500':
          description: Erreur interne du serveur

  /internal/api/v1/orders/{id}:
    get:
      tags:
        - "Orders"
      operationId: "getOrderById"
      parameters:
        - name: id
          in: path
          description: ID du commande à retrouver
          required: true
          schema:
            type: integer
            format: int64
      summary: Récupérer un commande à partir de son identifiant
      responses:
        '200':
          description: Succès. Renvoie la commande ayant l'identifiant concerné.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: ID invalide
        '404':
          description: Commande non retrouvable

    put:
      tags:
        - "Orders"
      operationId: "updateOrder"
      summary: Modifie une commande
      parameters:
        - name: id
          in: path
          description: ID de la commande à retrouver
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Commande modifiée avec succès
        '400':
          description: Mauvaise requête.
        '500':
          description: Erreur interne du serveur

    delete:
      tags:
        - "Orders"
      operationId: "deleteOrderById"
      parameters:
        - name: id
          in: path
          description: ID de la commande à supprimer
          required: true
          schema:
            type: integer
            format: int64
      summary: Supprimer une commande à partir de son identifiant
      responses:
        '400':
          description: ID invalide

components:
  schemas:
    Order:
      type: object
      required:
        - userId
        - productsIds
      properties:
        userId:
          type: integer
          format: int64
        productsIds:
          type: array
          items:
            type: integer
            format: int64